customer-support-platform/requirements.txt
# Customer Support Platform - Requirements File
# Generated from pyproject.toml for environments that don't support Poetry

# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain & LangGraph - AI/ML Framework
langchain==0.0.340
langchain-openai==0.0.2
langgraph==0.0.20
langchain-community==0.0.3

# Vector Stores
pinecone-client==2.2.4
weaviate-client==3.25.3
chromadb==0.4.18

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Monitoring & Observability
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-instrumentation-redis==0.42b0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
typer==0.9.0
rich==13.7.0
tenacity==8.2.3

# Message Queue
kafka-python==2.0.2
aiokafka==0.9.0

# NLP & ML
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
spacy==3.7.2
nltk==3.8.1

# Development Dependencies (Optional - Install with pip install -r requirements-dev.txt)
# pytest==7.4.3
# pytest-asyncio==0.23.2
# pytest-cov==4.1.0
# pytest-mock==3.12.0
# black==23.11.0
# isort==5.12.0
# flake8==6.1.0
# mypy==1.7.1
# pre-commit==3.6.0

# Production WSGI Server (Alternative to uvicorn)
# gunicorn==21.2.0

# Additional Production Dependencies
# sentry-sdk[fastapi]==1.38.0  # Error tracking
# newrelic==9.2.0  # Application monitoring
# datadog==0.48.0  # Monitoring and logging

# Streamlit UI
streamlit==1.40.2